refactor: extract touch to separate function
